+++ src/utils/formatCurrency.ts
